requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1
orjson==3.10.7
pydantic==2.9.2
pytest==8.2.2
pytest-asyncio==0.24.0
python-dotenv==1.0.1
numpy==1.26.4
